backend/requirements.txt
flask
flask-cors
numpy
scikit-fuzzy
scipy
networkx
requests
python-dotenv
gunicorn
